shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared code between orchestrator and doctor"
readme = "README.md"
authors = [
    { email = "luca.samore.01@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.4.7",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[project.scripts]
shared = "shared:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"